requirements.txt
streamlit
pandas
tabula-py
openpyx1